# feat: add stats option to analyze command for performance insights