@matthisholleville 2024-08-26
Linked to this issue https://github.com/k8sgpt-ai/k8sgpt/pull/1237