@AlexsJones 2024-08-26
If we do `-o json -s` what will happen, will it add a field into the json object?