@matthisholleville 2024-08-26
When this option is selected, the results are not displayed. Only the statistics are shown. If we try to display both, in the case of an analysis with many results, it becomes very difficult to visualize the statistics.