@AlexsJones 2024-08-26
> The current implementation does not support json as an output.

Might be worth considering if we want to show the results and time taken