@matthisholleville 2024-08-26
@AlexsJones stats implementation is global to analyzer. The results display the errors reported by an analyzer. In my opinion, these are two different notions. We could also display the stats in json format, if that's useful.